server setup

+++ server/modules/categories/category.controller.js
@@ -1,7 +1,9 @@
+const slugify = require("slugify");
 const Model = require("./category.model");
+const productModel = require("../products/product.model");
 
 const slugGenerator = (payload) => {
-  return payload.toLowerCase().replace(" ", "-");
+  return slugify(payload, { lower: true });
 };
 
 const create = async (payload) => {
@@ -14,6 +16,7 @@ const create = async (payload) => {
 };
 
 const list = async (limit = 10, page = 1, search) => {
+  page = page < 1 ? 1 : page;
   const query = [];
   if (search?.name) {
     query.push({
@@ -57,11 +60,6 @@ const list = async (limit = 10, page = 1, search) => {
         total: 1,
         data: 1,
       },
-    },
-    {
-      $project: {
-        "data.password": 0,
-      },
     }
   );
   const result = await Model.aggregate(query);
@@ -77,14 +75,21 @@ const getById = (id) => {
   return Model.findOne({ _id: id });
 };
 
-const updateById = (id, payload) => {
+const updateById = async (id, payload) => {
   if (payload.name) {
     payload.slug = slugGenerator(payload.name);
+    const isAvailable = await Model.findOne({ slug: payload.slug });
+    if (isAvailable) throw new Error("Category name is already in use");
   }
   return Model.findOneAndUpdate({ _id: id }, payload, { new: true });
 };
 
-const remove = (id) => {
+const remove = async (id) => {
+  const product = await productModel.findOne({ category: id });
+  if (product)
+    throw new Error(
+      `Remove category from product name ${product.name} to continue`
+    );
   return Model.deleteOne({ _id: id });
 };
